Create Seperate Module For Reset Functions
